# [libzstd] Fix FORCE_INLINE macro