* Pull compiler specific macros into `compiler.h`, which will work for zstd and fse.
* Split into `FORCE_INLINE_TEMPLATE` and `HINT_INLINE`. For gcc-4.8 and gcc-4.9 removing the `always_inline` attribute from `HINT_INLINE` gave 5-10 MB/s more decompression speed. Removing it from clang reduced decompression speed by 50 MB/s.

## Benchmarks

I ran on level 1 and 2 on silesia with `-b1e2i10`. I've listed the level 1 results blow, the level 2 results are consistent.

### Entropy Compression

| Compiler  | Before (MB/s) | After (MB/s) |
|-----------|---------------|--------------|
| gcc-5     |         438.6 |        438.6 |
| gcc-6     |         426.8 |        426.8 |
| gcc-7     |         438.6 |        441.6 |
| clang-3.8 |         444.7 |        444.7 |
| clang-3.9 |         429.7 |        426.8 |


### Devserver Compression

| Compiler  | Before (MB/s) | After (MB/s) |
|-----------|---------------|--------------|
| gcc-4.8   |         196.3 |        207.8 |
| gcc-4.9   |         191.0 |        207.8 |
| gcc-5     |         212.0 |        212.0 |
| gcc-7     |         209.9 |        212.0 |
| clang-5.0 |         212.0 |        209.9 |

### Entropy Decompression

| Compiler  | Before (MB/s) | After (MB/s) |
|-----------|---------------|--------------|
| gcc-5     |        1115.7 |       1115.7 |
| gcc-6     |        1106.0 |       1106.0 |
| gcc-7     |        1125.6 |       1125.6 |
| clang-3.8 |         883.3 |        883.3 |
| clang-3.9 |        1029.1 |       1029.1 |


### Devserver Decompression

| Compiler  | Before (MB/s) | After (MB/s) |
|-----------|---------------|--------------|
| gcc-4.8   |         553.0 |        557.9 |
| gcc-4.9   |         578.2 |        578.2 |
| gcc-5     |         543.6 |        543.6 |
| gcc-7     |         611.5 |        611.5 |
| clang-5.0 |         500.8 |        504.7 |