@Cyan4973 2017-08-15
Nice work @terrelln, it's not easy to get this setting right across so many target compiler versions.